test: get snapshot tests working
There's an issue with snapshot tests and the presence of "${}", backticks "`", or backslashes "\".
To get tests working, the markdown file being used for tests was modified to remove all instances of
those strings which cause the error. This of course is not ideal as we're not testing all markdown
rendering as we should, but at least now we have something that will catch some cases.

fix #99